refactor(utils): move util to util file
