+++ packages/utils/src/seconds-to-time/util.ts
@@ -1,0 +1,14 @@
+/**
+ * Приводит секунды к формату `hh:mm:ss` и возвращает объект с ними.
+ */
+export function secondsToTime(seconds: number) {
+    const h = Math.floor(seconds / 60 / 60);
+    const m = Math.floor(seconds / 60) % 60;
+    const s = seconds - h * 3600 - m * 60;
+
+    return {
+        hours: `${h}`.padStart(2, '0'),
+        minutes: `${m}`.padStart(2, '0'),
+        seconds: `${s}`.padStart(2, '0'),
+    };
+}
